Merge Sort - 1.2
The merge sort algorithm I have build here works only
for arrays with size of base 2. If odd numbered arrays
are generated in any time of the recursion (except when
the length of the arrays is 1) the entire algorithm fails
I have been unable to identify the problem with the code
upon some further researching I have found an algorithm
which takes in the beginning index and ending index of
the array which needs to be sorted in the mergeSort()
function. Thus I am creating a new branch, to deal with
this idea.