Merge Sort - 1.4
There was a problem in the logic of the mergeSort()
function as well as the problem in the variables of merge
wher the variables n and m were mixed up leading to
problems that went undetected. But these have been resolved now and the merge-sort works flawlessly (as fas as I know)
Hopefullt, no more problems will arise.